style(bounds): split budget.rs's inline tests into budget_tests.rs (#304)
This repo's change-risk gate (change_risk/signals.rs::is_test_file)
only recognizes a companion test file by path convention -- tests/**,
test_*/*_test(s).*, or a sibling foo_tests.rs -- not an inline
#[cfg(test)] mod tests {} block in the same file. budget.rs's 25 tests
were inline, so the gate scored this PR "source changed, no tests
touched" (the largest single weight in its formula) and blocked the
PR at 83/100 against an 80 threshold, even though every code path is
covered.

No behavior change: same 25 tests, moved verbatim to budget_tests.rs
and wired in via #[path], matching this repo's existing convention
(see artifacts.rs -> artifacts_tests.rs, change_risk/tests.rs, etc.).

Refs #157 #304.
